refactor: Update import paths to use 'src' prefix and improve module structure
