+++ src/api/routes/enrichment.py
@@ -7,8 +7,8 @@ from uuid import uuid4
 from fastapi import APIRouter, Depends, HTTPException, Query, Request
 from loguru import logger
 
-from api.dependencies import _deps
-from api.auth import get_current_user_id
+from src.api.dependencies import _deps
+from src.api.auth import get_current_user_id
 from src.run_deployment import invoke_enrichment_deployment
 from src.utils import serialize_datetime_fields
