+++ src/app.py
@@ -13,10 +13,10 @@ from loguru import logger
 
 from src.config import Config, create_dependencies
 from src.logging_config import setup_logging
-from src.socketio_instance import sio 
-from services.status_tracker import StatusTracker
-from api import router
-from api.dependencies import init_deps
+from src.socketio_instance import sio
+from src.services.status_tracker import StatusTracker
+from src.api import router
+from src.api.dependencies import init_deps
