+++ src/cravat_server.py
@@ -1,4 +1,12 @@
+import sys
+from pathlib import Path
+
 from flask import Flask, request, jsonify
+
+_repo_root = Path(__file__).resolve().parent.parent
+if str(_repo_root) not in sys.path:
+    sys.path.insert(0, str(_repo_root))
+
 from scripts.inference_util import analyze_coding_effect
 
 app = Flask(__name__)
