+++ src/flows/__init__.py
@@ -1,3 +1,3 @@
-from flows.hypothesis import hypothesis_flow
-from flows.enrichment import enrichment_flow, child_enrichment_batch_flow
-from flows.analysis import analysis_pipeline_flow
+from src.flows.hypothesis import hypothesis_flow
+from src.flows.enrichment import enrichment_flow, child_enrichment_batch_flow
+from src.flows.analysis import analysis_pipeline_flow
